requirements.txt
streamlit
opencv-python-headless
numpy
tensorflow-cpu
Pillow
h5py